FIX GENE.ONTO.ROOT can be empty
This patch fixes building the topology graph. Instead
of always relying on GENE.ONTO.ROOT to assess whether
the node being added to the graph is the root node,
it now either looks at GENE.ONTO.ROOT to check the
name or, if the name is empty, it checks whether the
node has adjacent nodes.

I've tested on a personal project that it reproduces
the same results as before.

closes #7